Very messy implementation of model into live feed.
